add .top code to module
