+++ AFMReader/top.py
@@ -1,4 +1,4 @@
-"""For decoding and loading .stp AFM file format into Python Numpy arrays."""
+"""For decoding and loading .top AFM file format into Python Numpy arrays."""
 
 from __future__ import annotations
 from pathlib import Path
@@ -7,22 +7,23 @@ import re
 import numpy as np
 
 from AFMReader.logging import logger
-from AFMReader.io import read_double
+from AFMReader.io import read_int16
 
 logger.enable(__package__)
 
 
 # pylint: disable=too-many-locals
-def load_stp(  # noqa: C901 (ignore too complex)
+# pylint: disable=too-many-statements
+def load_top(  # noqa: C901 (ignore too complex)
     file_path: Path | str, header_encoding: str = "latin-1"
 ) -> tuple[np.ndarray, float]:
     """
-    Load image from STP files.
+    Load image from TOP files.
 
     Parameters
     ----------
     file_path : Path | str
-        Path to the .stp file.
+        Path to the .top file.
     header_encoding : str
         Encoding to use for the header of the file. Default is ''latin-1''.
 
@@ -83,18 +84,26 @@ def load_stp(  # noqa: C901 (ignore too complex)
                     f"[{filename}] : X scan size (nm) does not equal Y scan size (nm) ({x_real_size}, {y_real_size})"
                     "we don't currently support non-square images."
                 )
+            Zrange_match = re.search(r"Z Amplitude: (\d+)", header_decoded)
+            if Zrange_match is None:
+                raise ValueError(f"[{filename}] : 'Z Amplitude' not found in file header.")
+            Zrange = float(Zrange_match.group(1))
 
             # Calculate pixel to nm scaling
             pixel_to_nm_scaling = x_real_size / cols
 
-            # Read R x C matrix of doubles
+            # Read R x C matrix of int16s
             image_list = []
             for _ in range(rows):
                 row = []
                 for _ in range(cols):
-                    row.append(read_double(open_file))
+                    row.append(read_int16(open_file))
                 image_list.append(row)
             image = np.array(image_list)
+            image_int = np.array(image_list)
+            Zmin = 0  # nm
+            norm_image = (image_int - np.min(image_int)) / (np.max(image_int) - np.min(image_int))
+            image = norm_image * (Zrange - Zmin) + Zmin
 
     except FileNotFoundError as e:
         logger.error(f"[{filename}] : File not found : {file_path}")
